format code, and fixed error at UNIX platform 2025-05-24  MerlotRain  <merlotrain@gmail.com>
    * include/dwg/CadObjectCollection.h:
    * include/dwg/CadSystemVariables_p.h:
    * include/dwg/Color.h:
    * include/dwg/SeqendCollection.h:
    * include/dwg/entities/Mesh.h:
    * include/dwg/entities/MultiLeader.h:
    * include/dwg/entities/PolyLine.h:
    * include/dwg/entities/Seqend.h:
    * include/dwg/io/dwg/DwgDocumentBuilder_p.h:
    * include/dwg/io/dwg/readers/DwgStreamReaderBase_p.h:
    * include/dwg/io/dwg/readers/DwgSummaryInfoReader_p.h:
    * include/dwg/io/dwg/readers/IDwgStreamReader_p.h:
    * include/dwg/io/template/CadBlockRecordTemplate_p.h:
    * include/dwg/io/template/CadBlockVisibilityParameterTemplate_p.h:
    * include/dwg/io/template/CadEvaluationGraphTemplate_p.h:
    * include/dwg/io/template/CadInsertTemplate_p.h:
    * include/dwg/io/template/CadMLeaderTemplate_p.h:
    * include/dwg/io/template/CadPolyLineTemplate_p.h:
    * include/dwg/io/template/CadTemplate_p.h:
    * include/dwg/objects/evaluations/Block1PtParameter.h:
    * include/dwg/objects/evaluations/Block2PtParameter.h:
    * include/dwg/objects/evaluations/EvaluationGraph.h:
    * include/dwg/tables/TextStyle.h:
    * include/dwg/utils/DateTime.h:
    * include/dwg/utils/DwgVariant.h:
    * include/dwg/utils/QFlags.h:
    * src/CadSystemVariables.cpp:
    * src/CadUtils.cpp:
    * src/entities/Mesh.cpp:
    * src/entities/Vertex.cpp:
    * src/header/CadHeader.cpp:
    * src/io/dwg/CRC32StreamHandler.cpp:
    * src/io/dwg/DwgReader.cpp:
    * src/io/dwg/readers/DwgStreamReaderBase.cpp:
    * src/io/dxf/readers/DxfBinaryReader.cpp:
    * src/io/dxf/readers/DxfStreamReaderBase.cpp:
    * src/io/dxf/writers/DxfHeaderSectionWriter.cpp:
    * src/io/template/CadAttributeTemplate.cpp:
    * src/io/template/CadEntityTemplate.cpp:
    * src/io/template/CadPolyfaceMeshTemplate.cpp:
    * src/io/template/CadTemplate.cpp:
    * src/objects/GeoData.cpp:
    * src/utils/DwgVariant.cpp:

+++ include/dwg/tables/TextStyle.h
@@ -81,7 +81,7 @@ private:
     double _lastHeight = 0.0;
     double _obliqueAngle = 0.0;
     TextMirrorFlag _mirrorFlag = TextMirrorFlag::None;
-    FontFlags _trueType = (int) FontFlag::Regular;
+    FontFlags _trueType = FontFlag::Regular;
 };
 
 }// namespace dwg
